🚚 Move settings-controls into new subfolder
