+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Controls/index.js
@@ -1,0 +1,5 @@
+export { default as ControlStaticValue } from './ControlStaticValue';
+export { default as ControlTextInput } from './ControlTextInput';
+export { default as ControlToggleButton } from './ControlToggleButton';
+export { default as ControlButton } from './ControlButton';
+export { default as ControlRadioGroup } from './ControlRadioGroup';
