+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/SettingsBlocks/index.js
@@ -1,9 +1,5 @@
 export { default as SelectSettingsBlock } from './SelectSettingsBlock';
 export { default as PaymentMethodsBlock } from './PaymentMethodsBlock';
 export { default as PaymentMethodItemBlock } from './PaymentMethodItemBlock';
-
-export { default as ControlStaticValue } from './ControlStaticValue';
-export { default as ControlTextInput } from './ControlTextInput';
-export { default as ControlToggleButton } from './ControlToggleButton';
-export { default as ControlButton } from './ControlButton';
-export { default as ControlRadioGroup } from './ControlRadioGroup';
+export { default as TodoSettingsBlock } from './TodoSettingsBlock';
+export { default as FeatureSettingsBlock } from './FeatureSettingsBlock';
