+++ modules/ppcp-settings/resources/js/Components/Screens/Overview/TabSettingsElements/Blocks/ConnectionDetails.js
@@ -4,7 +4,7 @@ import { Button } from '@wordpress/components';
 import {
 	ControlTextInput,
 	ControlRadioGroup,
-} from '../../../../ReusableComponents/SettingsBlocks';
+} from '../../../../ReusableComponents/Controls';
 import Accordion from '../../../../ReusableComponents/AccordionSection';
 import SettingsBlock from '../../../../ReusableComponents/SettingsBlock';
