+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Settings/Blocks/InvoicePrefix.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n';
 
-import { ControlTextInput } from '../../../../../ReusableComponents/SettingsBlocks';
+import { ControlTextInput } from '../../../../../ReusableComponents/Controls';
 import { SettingsHooks } from '../../../../../../data';
 import SettingsBlock from '../../../../../ReusableComponents/SettingsBlock';
