+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Components/Settings/Blocks/OrderIntent.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n';
 
-import { ControlToggleButton } from '../../../../../ReusableComponents/SettingsBlocks';
+import { ControlToggleButton } from '../../../../../ReusableComponents/Controls';
 import SettingsBlock from '../../../../../ReusableComponents/SettingsBlock';
 import { SettingsHooks } from '../../../../../../data';
